+++ R/load_notes.R
@@ -2,7 +2,7 @@
 #' @export
 #'
 #' @description Loads documents information into the R environment, which are:
-#' \itemize{
+#' \describe{
 #' \item{Cardiology: }{"car"}
 #' \item{Discharge: }{"dis"}
 #' \item{Endoscopy: }{"end"}