+++ tests/testthat/test-convert_notes.R
@@ -38,7 +38,7 @@ test_that("convert_dia run using sequential and parallel loading returns same re
 
 ## Compare loaded data with legacy data -----
 expect_true({
-  date_cols <- colnames(c_s)[which(as.vector(c_s[,lapply(.SD, class)])[1,] == "POSIXct")]
+  date_cols <- colnames(c_s)[which(as.vector(c_s[,lapply(.SD, class)][1,]) == "POSIXct")]
   suppressWarnings(c_s[,(date_cols):= lapply(.SD, as.character), .SDcols = date_cols])
   c_s[is.na(c_s)] <- ""
   #c_s[] <- lapply(c_s, gsub, pattern = '"', replacement = '')