+++ tests/testthat/test-load_lab.R
@@ -29,7 +29,7 @@ test_that("load_lab run using sequential and parallel loading returns same resul
 ## Compare loaded data with legacy data -----
 ### Convert dates to text and remove NAs
 expect_true({
-  date_cols <- colnames(d_s)[which(as.vector(d_s[,lapply(.SD, class)])[1,] == "POSIXct")]
+  date_cols <- colnames(d_s)[which(as.vector(d_s[,lapply(.SD, class)][1,]) == "POSIXct")]
   suppressWarnings(d_s[,(date_cols):= lapply(.SD, as.character), .SDcols = date_cols])
   d_s[is.na(d_s)] <- ""
   #d_s[] <- lapply(d_s, gsub, pattern = '"', replacement = '')